Backendd/requirements.txt
gradio
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
PyPDF2
chromadb
